+++ src/main/java/com/example/backend_auth0/data/repository/PacienteRepository.java
@@ -6,7 +6,7 @@ import com.example.backend_auth0.data.entities.Paciente;
 import com.example.backend_auth0.data.entities.Usuario;
 import com.example.backend_auth0.data.repository.base.BaseRepository;
 
-public interface PacienteRepository extends BaseRepository<Paciente, Integer> {
+public interface PacienteRepository extends BaseRepository<Paciente, Long> {
 
     Optional<Paciente> findByUsuario(Usuario usuario);
 }
