+++ src/main/java/com/example/backend_auth0/data/repository/TurnoRepository.java
@@ -3,5 +3,5 @@ package com.example.backend_auth0.data.repository;
 import com.example.backend_auth0.data.entities.Turno;
 import com.example.backend_auth0.data.repository.base.BaseRepository;
 
-public interface TurnoRepository extends BaseRepository<Turno, Integer> {
+public interface TurnoRepository extends BaseRepository<Turno, Long> {
 }
