+++ src/main/java/com/example/backend_auth0/data/repository/UsuarioRepository.java
@@ -4,6 +4,6 @@ import com.example.backend_auth0.data.entities.Usuario;
 import com.example.backend_auth0.data.repository.base.BaseRepository;
 import java.util.Optional;
 
-public interface UsuarioRepository extends BaseRepository<Usuario, Integer> {
+public interface UsuarioRepository extends BaseRepository<Usuario, Long> {
     Optional<Usuario> findByAuth0Id(String auth0Id);
 }
